Make it compatible again with go < 1.5
